\name{gnlm}
\title{Generalized Nonlinear Regression Library}
\description{
\code{\link{fit.dist}}	Fit Probability Distributions to Frequency Data

\code{\link{fmr}}		Generalized Nonlinear Regression with Finite Mixture

\code{\link{gnlr}}		Generalized Nonlinear Regression

\code{\link{gnlr3}}		Generalized Nonlinear Regression (3 Parameter Distributions)

\code{\link{nlr}}		Nonlinear Regression

\code{\link{nordr}}		Nonlinear Ordinal Regression

\code{\link{rs2}}		Two-factor Nonlinear Response Surface

\code{\link{rs3}}		Three-factor Nonlinear Response Surface
}
